fix locale-is-not-valid error.
This was caused because the ShowAllRegistrantsPdfJob was setting the available_locales.
Normal jobs (like Devise mailers) don't set them...and so 2 different apartments could conflict
if the ShowAllRegistrantsPdfJob set a restricted set of available_locales, and then
a different apartment tried to run a job which requires a locale which isn't in that set.