Remove use of placeholders for clarity.
the problem with using placeholders is that as soon as you fill
in the fields, it's not possible to refer back to the placeholder to ensure
that you did it correctly.

As a result, the information provided in the placeholder is nearly useless
as soon as the user starts typing.

As an alternative, we use 'hints' which appear beneath the input box,
and don't disappear