Fix form-submission. Clicks the button because we care which button is clicked.
If we only submit the form, it doesn't create a single competitor, because it's
looking for the text of the button which was clicked.

Also fixes the re-display of the form when form-submission fails.
(it was showing that the failed ids were added to a competitor, when they were not)